Change a declaration name to be more descriptive
Generally the excl::struct-by-value declaration lodges two index
values into an argument definition, based on what kind of struct
argument it is and where it is in the argument list.  Usually these
are an int index and a float index.  But the panasonic (and, I
suspect, other systems), want pass-by-value structs that are larger
than two words to be passed on the stack (and not in registers) so
I've renamed the second offset from float-offset to other-offset.
This is part of the panasonic porting effort.

Change-Id: I3a209137e31c3db643f09506b067dec8357726bc